+++ src/GameState/GameStateManager.java
@@ -12,6 +12,7 @@ public class GameStateManager {
 
         currentState = MENUSTATE;
         gameStates.add(new MenuState(this));
+        gameStates.add(new Level1State(this));
 
     }
