+++ src/GameState/Level1State.java
@@ -1,0 +1,36 @@
+package GameState;
+
+import TileMap.TileMap;
+
+import java.awt.*;
+
+public class Level1State extends GameState{
+    private TileMap tileMap;
+    public Level1State(GameStateManager gsm) {
+        this.gsm = gsm;
+        init();
+    }
+
+    public void init() {
+    }
+
+
+    public void update() {
+
+    }
+
+
+    public void draw(Graphics2D g) {
+
+    }
+
+
+    public void keyPressed(int k) {
+
+    }
+
+
+    public void keyReleased(int k) {
+
+    }
+}
