+++ src/GameState/MenuState.java
@@ -63,7 +63,7 @@ public class MenuState extends GameState {
 
     private void select() {
         if (currentChoice == 0) {
-            //start
+
         }
         if (currentChoice == 1) {
             //help
